refactor: ErrorAlert 타입 정의 및 AddKeyboardNotification 프로토콜 파일 분리
