+++ Diary/Controller/AddDiaryViewController.swift
@@ -7,39 +7,6 @@
 
 import UIKit
 
-@objc protocol AddKeyboardNotification {
-    @objc func keyboardWillShow(notification: NSNotification)
-    @objc func keyboardWillHide(notification: NSNotification)
-}
-
-extension AddKeyboardNotification {
-    func setKeyboardObserver() {
-        NotificationCenter.default.addObserver(self,
-                                               selector: #selector(self.keyboardWillShow),
-                                               name: UIResponder.keyboardWillShowNotification,
-                                               object: nil)
-        
-        NotificationCenter.default.addObserver(self,
-                                               selector: #selector(self.keyboardWillHide),
-                                               name: UIResponder.keyboardWillHideNotification,
-                                               object: nil)
-    }
-    
-    func getKeyboardHeight(from notification: NSNotification) -> CGFloat? {
-        guard let userInfo: NSDictionary = notification.userInfo as? NSDictionary else {
-            return nil
-        }
-        
-        guard let keyboardFrame = userInfo.value(forKey: UIResponder.keyboardFrameEndUserInfoKey) as? NSValue else {
-            return nil
-        }
-        
-        let keyboardRectangle = keyboardFrame.cgRectValue
-        
-        return keyboardRectangle.height
-    }
-}
-
 final class AddDiaryViewController: UIViewController, AddKeyboardNotification {
     private let addDiaryView = AddDiaryView()
